+++ src/main/java/main/algorithms/impls/v1/download/PeersToPiecesMapperImpl.java
@@ -1,4 +1,4 @@
-package main.algorithms.impl.v1.download;
+package main.algorithms.impls.v1.download;
 
 import main.algorithms.PeersToPiecesMapper;
 import main.peer.Link;
